requirements.txt
# Project dependencies
gradio==5.16.0
haystack-ai==2.10.0
trafilatura==2.0.0
openai==1.63.2
python-dotenv==1.0.1
# Development dependencies
doit==0.36.0
pylint==3.3.4
mdformat==0.7.22
black==25.1.0
isort==6.0.0